5. Bitcamp-Study-Python(ML)/requirements.txt
aniso8601==8.0.0
beautifulsoup4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
folium==0.12.1
googlemaps==2.5.1
html5lib
icecream==2.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JPype1
jsonify==0.5
jupyter==1.0.0
konlpy==0.6.0
lxml
MarkupSafe==1.1.1
mccabe
mysql-connector
nltk
numpy
pandas
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5.0.0
python-dateutil==2.8.2
pytz==2020.1
requests==2.24.0
seaborn
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
mysqlclient
tensorflow==2.7.0
tensorflow-gpu==2.7.0
tqdm
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
torch
pytorch==1.10.1
tensorflow-datasets
tensorflow-hub
tensorflow-metadata
django==3.2.5
djangorestframework
scipy==1.7.3
keras-gpu==2.6.0
konlpy==0.6.0
